move smid to mod, and perf improvements
