+++ src/object/mod.rs
@@ -1,7 +1,6 @@
 pub mod list;
-pub mod simd;
 pub mod string;
 
 pub use list::{create_list, create_list_empty, init_list_type, list_set_item_transfer};
-pub use simd::convert_pystring;
+pub use crate::simd::convert_pystring;
 pub use string::{create_fast_string, init_string_type};
